# db: copy user-provided bounds and optimize unchanged bounds