@erikgrinaker 2022-04-29
I ran some benchmarks for scans and gets, using a single batch that clones an iterator and calls `SetOptions()` on it for every operation. Gets use `Prefix: true` and `nil` bounds, scans use random upper/lower bounds for every operation. I compared this PR with its immediate parent, on top of cockroachdb/cockroach#79993.

```
name                                                 old time/op    new time/op    delta
MVCCScan_Pebble/rows=1/versions=1/valueSize=8-24       3.37µs ± 2%    3.31µs ± 1%   -1.54%  (p=0.000 n=19+19)
MVCCScan_Pebble/rows=10/versions=1/valueSize=8-24      6.56µs ± 2%    6.53µs ± 1%   -0.54%  (p=0.013 n=18+18)
MVCCScan_Pebble/rows=100/versions=1/valueSize=8-24     30.3µs ± 2%    30.5µs ± 2%   +0.72%  (p=0.015 n=20+20)
MVCCScan_Pebble/rows=1000/versions=1/valueSize=8-24     257µs ± 1%     257µs ± 2%     ~     (p=0.517 n=17+20)
MVCCGet_Pebble/versions=1/valueSize=8-24               2.68µs ± 1%    2.91µs ± 1%   +8.64%  (p=0.000 n=20+16)
MVCCGet_Pebble/versions=10/valueSize=8-24              4.07µs ± 2%    4.39µs ± 3%   +7.94%  (p=0.000 n=18+19)
MVCCGet_Pebble/versions=100/valueSize=8-24             10.2µs ± 4%    11.4µs ± 5%  +11.46%  (p=0.000 n=20+20)
```

I would expect the copy overhead to show up in the scan benchmarks, but it appears to be negligible, so I don't think the copy is a problem in itself.

For gets, I looked at profile diffs, and it seemed like it was mostly down to `SeekGE` overhead. That's likely because this affects the lower-level seek optimization, and should be orthogonal to the copies themselves. Profile diff below:

<img width="307" alt="Screenshot 2022-04-29 at 17 24 11" src="https://user-images.githubusercontent.com/644420/165975334-6dd344a8-f7c5-4081-9f4b-b8b5fe88fb3b.png">